New feature: move a pawn
Enables moving the pawn at a2 to a3. When onDrag event is thrown for that pawn, it creates a list of possible moves for that piece. The dropzones are then designated via React. When onDrop event is thrown, the pawn is moved in state. React then renders the new state.